Extract function to `threshold-ts` lib utils
Move function that prefixes the output script with `0x` and its own
length to `threshold-ts` lib utils.
