+++ src/hooks/tbtc/useFetchRedemptionDetails.ts
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react"
 import { useThreshold } from "../../contexts/ThresholdContext"
+import { prependScriptPubKeyByLength } from "../../threshold-ts/utils"
 import { useGetBlock } from "../../web3/hooks"
 
 interface RedemptionDetails {
@@ -161,7 +162,7 @@ export const useFetchRedemptionDetails = (
 
           for (const { scriptPubKey } of outputs) {
             if (
-              toPrefixedRawRedeemerOutputScript(scriptPubKey.toString()) !==
+              prependScriptPubKeyByLength(scriptPubKey.toString()) !==
               redemptionRequest.redeemerOutputScript
             )
               continue
@@ -212,14 +213,4 @@ export const useFetchRedemptionDetails = (
   ])
 
   return { isFetching, data: redemptionData, error }
-}
-
-const toPrefixedRawRedeemerOutputScript = (scriptPubKey: string) => {
-  const rawRedeemerOutputScript = Buffer.from(scriptPubKey.toString(), "hex")
-
-  // Prefix the output script bytes buffer with 0x and its own length.
-  return `0x${Buffer.concat([
-    Buffer.from([rawRedeemerOutputScript.length]),
-    rawRedeemerOutputScript,
-  ]).toString("hex")}`
 }
